==> Data_Cleaning_Only/0 Explanation of Data Set.ipynb <==
# Explanation of Data Set

Our data set was obtained by running the Youtube Shorts Data from Apify. We scraped starting on 3/1 over the course of several days. We scraped all posts from at least a year back from a list of influncers. We chose as many influencers as we could with more than 1,000 subscribers who focused on makeup and skincare and who were primarily concerned with product reviews. Our sample is not a perfect random sample (which would be impossible to obtain without access to a full list of youtubers in the beauty niche) but should include the majority of all major influencers from this niche. We ended up with 155 names.

One thing to note about the date data. The time zone was not included when we pulled the data set. To figure out the time zone for the data we posted a test short from a test channel at 9am EST. We then pulled the shorts from that channel and saw that the scraper listed the time of posting as 9am EST. This confirms that the date data in our data set is in EST. Note that everyone who used the scraper to pull the data was also located in the same time zone so local differences did not affect the data. 

The exact influencers we pulled from are as follows: dy
auty 
Kerry K 
slayyys
MissDarcei
@AnifeToupeva
Morgan @Glolabmorgan
aBankson
RudiBerry
Melissa Alatorre
urusrex 
Nikkia Joy (also a brand owner)


  Angelica Nyqvist<br>
  Morgan Turner <br>
  Theresa is Dead
<br>  Jen Phelps
<br>Rachel Palmieri
<br>Elle Chu
<br>Andrea Matilliano
<br>Makeup.Just.For.Fun
<br>Battybean
<br>Makeup By Nikki La Rose
<br>Pampered Wolf
<br>Patty Alonso
<br>State of Kait
<br>Kelly Gooch 
<br>Hautemess tom
<br>Badtothebrow
<br>Sarah Rose
<br>Allie Glines
<br>Stephanie Marie
<br>Lauren Mae Beauty
<br>Glitter Fallout
<br>Juicyjas
<br>MakeupByCheryl
<br>Angela Bright
<br>Evangeline Molly 
<br>Risa Does Makeup
<br>Charlotte Holdcroft
<br>BabsBeauty
<br>Rainier Cramer
<br>Bekah Sun
<br>Julia Adams MUA
<br>Better Than Basic
<br>Ella Kinkley
<br>Lauren Lehnen
<br>CourtneyQuarantine
<br>Slashed Beauty
<br>Casey Holmes
<br>ThatsSoYin
<br>Taylor Margaret 
<br>Theoliviasaurusrex 
<br>Nikkia Joy
<br>NikkieTutorials
<br>Huda Kattan 
<br>Jackie Aina
<br>Michelle Phan
<br>KathleenLights
<br>Manny MUA
<br>Desi Perkins
<br>RawBeautyKristi
<br>Chloe Morello
<br>Patrick Starr
<br>Wayne Goss 
<br>RachhLoves
<br>Melissa Alatorre
<br>Sofiepavitt
<br>Alexgreysonskin
<br>Xmeloooody
<br>Lulukehlani
<br>Skincareanarchythepodcast
<br>NadiaVegaYT
<br>Thebellesophie
<br>Skincarebyaliza
<br>s0hyuna
<br>Valezrina
<br>Sarahpalmyra
<br>Bellbasaur
<br>Shereeneidriss
<br>AnnettesMakeupCorner
<br>KellyStrack
<br>KellyDriscoll
<br>Worldofcolorx
<br>Alishayh
<br>TomiwaRodia
<br>MichelleBali
<br>EleinaHinch
<br>Blueyoobin
<br>CYounginyou
<br>AliceintheRabbitHole
<br>Stationjaya
<br>Ltejuvu
<br>thebeautybreakdown
<br>Lindsey (teacher) @lindseyteaches
<br>Anaël Hector @anaelhector
<br>Aneeta Wilson @antaytay
<br>Esther Daze @estherdaze 
<br>Trivima Gupta @theglammyvibe
<br>Emily Noel @emilynoel83
<br>Anife @AnifeToupeva
<br>Morgan @Glolabmorgan
<br>NymaTang
<br>TaylorAnise
<br>Ronkeraji
<br>Monica.raviii
<br>Sequerstin
<br>Jasadetunji
<br>Themyraaakn
<br>MorganTurnerMakeup 
<br>Ohemaabonsu
<br>Jazlmao
<br>Emmaaurelia
<br>StateofKait
<br>Christxiee
<br>Golloriaa
<br>KristenMarie
<br>SydneyPurl
<br>Tenishaward
<br>DermAngelo
<br>Salonislayyys
<br>MissDarcei
<br>AleJay
<br>SamanthaMarchxo
<br>AmandaZ
<br>j.uuuuudy
<br>VictoriaLyn
<br>HeatherAustinmakeup
<br>Kackie
<br>Julia.P 
<br>TaniaBWells
<br>Katelynhousemakeup
<br>Lifebyjessica1 
<br>KarenHarrismakeup
<br>JodiMannes
<br>LeahJanae 
<br>SofiaSeesBeauty 
<br>Kerry_K 
<br>Reetikab
<br>SPFDiary
<br>Beautykat3402
<br>Wjamc
<br>Jadeskinjoyce
<br>Nats_tutorials
<br>Ilincaw
<br>Lordthivii
<br>Panwithsophx
<br>Glamgio_nyc
<br>AylenPark
<br>Minseonk1m
<br>Meebebe_
<br>Zoerammie
<br>Leeskincare730
<br>MattRandon
<br>Jeenie.weenie
<br>Krystallee
<br>Kiyokochii
<br>Cherryantacidsx
<br>Nekoyanin
<br>AtarahMayhew
<br>FROARTISTRY
<br>StephiGlam
<br>Fxfreak
<br>ItsSisha
<br>TrinaMerryMakeu
<br>jonnybeauty9468pSkincare
<br>Slashedbeauty
<br>beautyforeverybody
