==> Data_Cleaning_Only/1 data_cleaning1.ipynb <==
In this notebook we do some basic data cleaning. Our data arrived in batches so we need to concatenate the data frames together into one data set. 
We then dropped a bunch of columns that did not contain anything of interest to us or contained garbage information. We also dropped any videos that may have been accidentally scraped twice. 

import os
import glob
import pandas as pd

#Define the folder containing CSV files
folder_path = r"C:\Users\arubi\Desktop\datascience2025\erdos-uncreatives\data\data_to_concatenate"

#Use glob to get CSV file paths
csv_files = glob.glob(os.path.join(folder_path, "*.csv"))

#read and concatenate all CSV files
df_list = [pd.read_csv(file) for file in csv_files]
concatenated_df = pd.concat(df_list, ignore_index=True)

#Save the concatenated CSV file
#output_file = os.path.join(folder_path, "concatenated_data.csv")
#concatenated_df.to_csv('/Users/gabrielaterrazas/Desktop/erdos-uncreatives-1/raw_data_read_only/data_cleaning/concatenated_data.csv', index=False)

print(concatenated_df)

 # Drop columns: 'aboutChannelInfo/channelAvatarUrl', 'aboutChannelInfo/channelBannerUrl' and 88 other columns
def clean_data(df):
    df = df.drop(columns=['aboutChannelInfo/channelAvatarUrl', 'aboutChannelInfo/channelBannerUrl', 'aboutChannelInfo/channelDescription', 'aboutChannelInfo/channelDescriptionLinks/0/text', 'aboutChannelInfo/channelDescriptionLinks/0/url', 'aboutChannelInfo/channelDescriptionLinks/1/text', 'aboutChannelInfo/channelDescriptionLinks/1/url', 'aboutChannelInfo/channelDescriptionLinks/2/text', 'aboutChannelInfo/channelDescriptionLinks/2/url', 'aboutChannelInfo/channelDescriptionLinks/3/text', 'aboutChannelInfo/channelDescriptionLinks/3/url', 'aboutChannelInfo/channelDescriptionLinks/4/text', 'aboutChannelInfo/channelDescriptionLinks/4/url', 'aboutChannelInfo/channelDescriptionLinks/5/text', 'aboutChannelInfo/channelDescriptionLinks/5/url', 'aboutChannelInfo/channelDescriptionLinks/6/text', 'aboutChannelInfo/channelDescriptionLinks/6/url', 'aboutChannelInfo/channelDescriptionLinks/7/text', 'aboutChannelInfo/channelDescriptionLinks/7/url', 'aboutChannelInfo/channelDescriptionLinks/8/text', 'aboutChannelInfo/channelDescriptionLinks/8/url', 'aboutChannelInfo/channelDescriptionLinks/9/text', 'aboutChannelInfo/channelDescriptionLinks/9/url', 'aboutChannelInfo/channelDescriptionLinks/10/text', 'aboutChannelInfo/channelDescriptionLinks/10/url', 'aboutChannelInfo/channelDescriptionLinks/11/text', 'aboutChannelInfo/channelDescriptionLinks/11/url', 'aboutChannelInfo/channelDescriptionLinks/12/text', 'aboutChannelInfo/channelDescriptionLinks/12/url', 'aboutChannelInfo/channelId', 'aboutChannelInfo/channelJoinedDate', 'aboutChannelInfo/channelLocation', 'aboutChannelInfo/channelName', 'aboutChannelInfo/channelTotalVideos', 'aboutChannelInfo/channelTotalViews', 'aboutChannelInfo/channelUrl', 'aboutChannelInfo/channelUsername', 'aboutChannelInfo/inputChannelUrl', 'aboutChannelInfo/isAgeRestricted', 'aboutChannelInfo/isChannelVerified', 'aboutChannelInfo/numberOfSubscribers', 'channelUrl', 'formats', 'channelAvatarUrl', 'channelBannerUrl', 'channelDescriptionLinks/0/text', 'channelDescriptionLinks/1/text', 'channelDescriptionLinks/2/text', 'channelDescriptionLinks/3/text', 'channelDescriptionLinks/4/text', 'channelDescriptionLinks/5/text', 'channelDescriptionLinks/6/text', 'channelDescriptionLinks/7/text', 'channelDescriptionLinks/8/text', 'channelDescriptionLinks/9/text', 'channelDescriptionLinks/10/text', 'channelDescriptionLinks/11/text', 'channelDescriptionLinks/12/text', 'channelId', 'channelName', 'comments', 'descriptionLinks', 'fromChannelListPage', 'fromYTUrl', 'inputChannelUrl', 'isMonetized', 'note', 'progressKey/label', 'progressKey/url', 'standardizedUrl', 'subtitles', 'thumbnailUrl', 'isAgeRestricted', 'input', 'type', 'channelDescriptionLinks/0/url', 'channelDescriptionLinks/1/url', 'channelDescriptionLinks/2/url', 'channelDescriptionLinks/3/url', 'channelDescriptionLinks/4/url', 'channelDescriptionLinks/5/url', 'channelDescriptionLinks/6/url', 'channelDescriptionLinks/7/url', 'channelDescriptionLinks/8/url', 'channelDescriptionLinks/9/url', 'channelDescriptionLinks/10/url', 'channelDescriptionLinks/11/url', 'channelDescriptionLinks/12/url'])
    return df

# Loaded variable 'df' from URI: /Users/gabrielaterrazas/Desktop/erdos-uncreatives-1/raw_data_read_only/data_cleaning/concatenated_data.csv
#df = pd.read_csv(r'/Users/gabrielaterrazas/Desktop/erdos-uncreatives-1/raw_data_read_only/data_cleaning/concatenated_data.csv')
df = concatenated_df
df_clean = clean_data(df.copy())


#Save the cleaned_data CSV file
#output_file = os.path.join(folder_path, "clean_data.csv")
#df_clean.to_csv('/Users/gabrielaterrazas/Desktop/erdos-uncreatives-1/raw_data_read_only/data_cleaning/clean_data.csv', index=False)



# Drop rows with missing and duplicated data in column: 'id' this column contains the unique identifier for each video
#if data is missing that means that the information does not pertain to a video, mostly is grabbed from the about page
# and if data is duplicated that means that the same video was scraped twice
def clean_data(df):
    df = df.dropna(subset=['id'])
    # Drop duplicate rows in column: 'id'
    df = df.drop_duplicates(subset=['id'])
    return df

# Loaded variable 'df' from URI: /Users/gabrielaterrazas/Desktop/erdos-uncreatives-1/raw_data_read_only/data_cleaning/clean_data.csv
#df = pd.read_csv(r'/Users/gabrielaterrazas/Desktop/erdos-uncreatives-1/raw_data_read_only/data_cleaning/clean_data.csv')

df = concatenated_df
df_clean = clean_data(df.copy())

#Save the new cleaned_data CSV file
#output_file = os.path.join(folder_path, "clean_data2.csv")
df_clean.to_csv("../data/clean_data2.csv")
