==> Data_Cleaning_Only/2 data_cleaning2.ipynb <==
In this notebook we continue data cleaning. We drop any columns that contained empty data which would be an error from the data scraping process. We convert the duration column into a numerical column that lists the duration of the short in seconds. We add a few post performance metrics such as a column that lists likes/subscriber. Finally, we make sure that our data set only contains shorts from 3/1/2024 to 3/1/2025 which is the time period of posts that we wanted to look at for our project. 

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

#importing our slightly cleaned data
#Changed to a relative directory so anyone can run it
df = pd.read_csv("../data/clean_data2.csv")
print(df.columns) #Investigating which columns remain after cleaning 

df.info()

df["text"].sample(10)
#THIS IS THE "DESCRIPTION" OF THE VIDEO. MANY USERS LEAVE THE DESCRIPTION BLANK!!! 
#We want this. 
#I am changing the NaN entries to empty strings. 
# This is an easier and more performant way of doing this
# Also doing this for channelDescription
df["text"] = df["text"].fillna("")
df["channelDescription"] = df["channelDescription"].fillna("")

df["text"].sample(10)

#Double-checking for non-null entries
#Let's drop Channel Location since that one is also often empty

# There's no real point in incrementing the name of the dataframe here

df = df.drop("channelLocation", axis = 1) 
df.info()

#Okay, now I see that some columns still have empty entries. Let's drop them. 

df = df.dropna().copy()
df.info()

#We are interested in engagement metrics. Obviously the more subscribers you have, the more views/likes/comments you will get. A post is performing well
#regardless of follower count if it has good "engagement", which is a metrics such as likes/subscriber. Here I am transforming raw likes/comments/views
#to engagement metrics. 

# Make sure to use underscores in column names. Spaces interfere with some functions such as pd.query()

df["likes_per_subscriber"] = df["likes"]/df["numberOfSubscribers"]
df["comments_per_subscriber"] = df["commentsCount"]/df["numberOfSubscribers"]
df["views_per_subscriber"] = df["viewCount"]/df["numberOfSubscribers"]


#Changing duration column from a string to an integer 
import datetime 

def convert_duration_to_int(item):
    item_as_datetime = datetime.datetime.strptime( item , "%H:%M:%S"  ) #converts string to a datetime object
    seconds = item_as_datetime.second + 60 * item_as_datetime.minute + 3600 * item_as_datetime.hour
    return seconds
    
df["duration_in_seconds"] = df["duration"].apply(convert_duration_to_int) 
df[["duration", "duration_in_seconds"]].sample(10) #Can confirm that column is correct now

#We should change our data so that none of the data comes after 3/1/2025 and none comes before 3/1/2024. 

#Okay, let's restrict our attention to those rows that came after October 15th.
#At the same time, we should change our data so that none of the data comes after 3/1. This is the date we first started collecting
#and after that date, data was scraped over the first few days. 

#Okay, time to continue to clean by removing data after 3/1/2025 and before 3/1/2024.
#First, we need to change the date column so that it is a datetime object and not a string

df["datetime"] = df["date"].apply(datetime.datetime.fromisoformat)
df["datetime_date"] = df["datetime"].dt.date.astype(str)
df["datetime_date"]

df = df.query("datetime_date < '2025-03-01'")
df = df.query("datetime_date > '2024-03-01'")

df.to_csv("../data/cleaned_data_correct_dates.csv")
